glassworm-protocol/requirements.txt
requests==2.33.1
PyGithub==2.9.0
